requirements.txt
astroid==3.3.8
cfgv==3.4.0
colorama==0.4.6
dill==0.3.9
distlib==0.3.9
filelock==3.17.0
greenlet==3.1.1
identify==2.6.6
iniconfig==2.0.0
isort==6.0.0
mccabe==0.7.0
mysql-connector-python==9.2.0
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.1.0
pylint==3.3.4
PyMySQL==1.1.1
pytest==8.3.4
python-dotenv==1.0.1
PyYAML==6.0.2
SQLAlchemy==2.0.37
tomlkit==0.13.2
typing_extensions==4.12.2
virtualenv==20.29.1
